+++ decidim-blogs/app/forms/decidim/blogs/admin/post_form.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+module Decidim
+  module Blogs
+    module Admin
+      # This class holds a Form to update pages from Decidim's admin panel.
+      class PostForm < Decidim::Form
+        include TranslatableAttributes
+
+        translatable_attribute :title, String
+        translatable_attribute :body, String
+
+        validates :title, translatable_presence: true
+        validates :body, translatable_presence: true
+      end
+    end
+  end
+end